DEV-74647 - Support for source code files

+++ src/helper.py
@@ -1,7 +1,9 @@
 import json
 import logging
+import shutil
 import urllib.parse
 import uuid
+from pathlib import Path
 from typing import List, Optional, Sequence, Union
 
 import requests
@@ -77,7 +79,7 @@ def generate_source_code(
     """
     Helper function that generates `SourceCode` object.
 
-    :param source_code_text: Text to use as source code
+    :param source_code_text: Text to use as source code or path of a file to be used
     :type source_code_text: str
     :param custom_lineage_config: Configuration object
     :type custom_lineage_config: CustomLineageConfig
@@ -88,10 +90,15 @@ def generate_source_code(
     :returns: SourceCode object constructed using the provided input
     :rtype: SourceCode
     """
-    # generate file name and create directory
-    file_name = f"{str(uuid.uuid4())}.txt"
-    with open(custom_lineage_config.source_code_directory_path / file_name, "w") as out_file:
-        out_file.write(source_code_text)
+    # in case of a file
+    if Path(source_code_text).is_file():
+        file_name = Path(source_code_text).name
+        shutil.copy(source_code_text, custom_lineage_config.source_code_directory_path / file_name)
+    else:
+        # generate file name
+        file_name = f"{str(uuid.uuid4())}.txt"
+        with open(custom_lineage_config.source_code_directory_path / file_name, "w") as out_file:
+            out_file.write(source_code_text)
     return SourceCode(
         path=f"{custom_lineage_config.source_code_directory_name}/{file_name}",
         highlights=highlights,
